[REFACTOR] 프로필 화면 레이어 분리 및 mock API 로그 추가
